+++ src/components/ResourceList/components/FilterControl/tests/FilterCreator.test.tsx
@@ -38,6 +38,21 @@ describe('<FilterCreator />', () => {
         label: 'Tagged with',
         type: FilterType.TextField,
       },
+      {
+        key: 'filterKey3',
+        label: 'Times used',
+        operatorText: [
+          {
+            optionLabel: 'less than',
+            key: 'times_used_max',
+          },
+          {
+            optionLabel: 'greater than',
+            key: 'times_used_min',
+          },
+        ],
+        type: FilterType.TextField,
+      },
     ],
     resourceName: {
       singular: 'Item',
@@ -149,6 +164,10 @@ describe('<FilterCreator />', () => {
           value: mockDefaultProps.filters[1].key,
           label: mockDefaultProps.filters[1].label,
         },
+        {
+          value: mockDefaultProps.filters[2].key,
+          label: mockDefaultProps.filters[2].label,
+        },
       ]);
     });
   });
@@ -164,7 +183,7 @@ describe('<FilterCreator />', () => {
       expect(wrapper.find(FilterValueSelector).exists()).toBe(false);
     });
 
-    it('updates FilterValueSelector when user select a filter key', () => {
+    it('updates FilterValueSelector when user selects a filter key', () => {
       const wrapper = mountWithAppProvider(
         <FilterCreator {...mockDefaultProps} />,
       );
@@ -178,7 +197,7 @@ describe('<FilterCreator />', () => {
       expect(wrapper.find(FilterValueSelector).prop('value')).toBeUndefined();
     });
 
-    it('updates value correctly when user select a filter value', () => {
+    it('updates value correctly when user selects a filter value', () => {
       const wrapper = mountWithAppProvider(
         <FilterCreator {...mockDefaultProps} />,
       );
@@ -188,6 +207,28 @@ describe('<FilterCreator />', () => {
       selectFilterValue(wrapper, 'Bundle');
 
       expect(wrapper.find(FilterValueSelector).prop('value')).toBe('Bundle');
+    });
+
+    it('updates FilterValueSelector when filter key is updated to existing operator key', () => {
+      const wrapper = mountWithAppProvider(
+        <FilterCreator {...mockDefaultProps} />,
+      );
+
+      const newOperatorKey = 'times_used_max';
+
+      activatePopover(wrapper);
+      selectFilterKey(wrapper, mockDefaultProps.filters[2].key);
+      selectFilterValue(wrapper, 'Bundle');
+
+      trigger(
+        wrapper.find(FilterValueSelector),
+        'onFilterKeyChange',
+        newOperatorKey,
+      );
+
+      expect(wrapper.find(FilterValueSelector).prop('filterKey')).toBe(
+        newOperatorKey,
+      );
     });
   });